test: Update names of helpers
