--- helpers/extractSelectedDisplayedValue.ts
@@ -1,9 +1,0 @@
-import { Locator } from "@playwright/test";
-
-export async function extractSelectedDisplayedValue(
-  selectedOption,
-): Promise<string> {
-  return selectedOption.evaluate(
-    (sel) => sel.options[sel.options.selectedIndex].textContent,
-  );
-}
